pyproject.toml
[project]
name = "cardtrader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.8",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.39.0",
]
